Extract site settings controls
